samples/web-app-sql-database/python/src/requirements.txt
Flask==3.1.0
azure-identity==1.25.1
pyodbc==5.3.0
gunicorn==20.1.0
python-dotenv==1.1.1
azure-keyvault-secrets